+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add possibility to configure DeploymentStrategy for Kafka Connect, Kafka Mirror Maker (1 and 2), and Kafka Bridge
 * Support passing metrics configuration as an external ConfigMap
 * Enable CORS configuration for Cruise Control
+* Add support for rolling individual Kafka or ZooKeeper pods through the Cluster Operator using an annotation
 
 ### Deprecations and removals
 * The `metrics` field in the Strimzi custom resources has been deprecated and will be removed in the future. For configuring metrics, use the new `metricsConfig` field and pass the configuration via ConfigMap.